+++ eon_timer/timers/gen3/model.py
@@ -3,7 +3,7 @@ from typing import override
 
 from eon_timer.util.enum import EnhancedEnum
 from eon_timer.util.injector import component
-from eon_timer.util.properties.property import Property, IntProperty
+from eon_timer.util.properties.property import Property, IntProperty, FloatProperty
 from eon_timer.util.properties.settings import Settings
 
 
@@ -17,7 +17,7 @@ class Gen3Model(Settings):
     mode = Property(Gen3Mode.STANDARD, value_type=str)
     pre_timer = IntProperty(5000)
     target_frame = IntProperty(1000)
-    calibration = IntProperty(0)
+    calibration = FloatProperty(0.0)
     frame_hit = IntProperty(0, transient=True)
 
     @property
